very tight zoom -- inspecting a few kHz -- was causing changes of scale in doZoom() due to floating point infelicities.  the frequency tape/grid would no longer line up with the FFT and waterfall.  this change to double is a quick workaround, but is not ideal. a quick workaround.

+++ core/src/gui/widgets/waterfall.h
@@ -103,11 +103,11 @@ namespace ImGui {
                 width = 524288;
             }
 
-            float factor = (float)width / (float)outWidth;
+            double factor = (float)width / (float)outWidth;
             std::cout << width << " " << outWidth << " " << factor << std::endl;
             float sFactor = ceilf(factor);
             float uFactor;
-            float id = offset;
+            double id = offset;
             float maxVal;
             int sId;
             for (int i = 0; i < outWidth; i++) {
